Remove some dependencies to gedit-debug
gedit-debug is not reusable for other projects, the section names are
hardcoded. So it can not go in the libgedit submodule.
So the code that will go in libgedit must not use gedit-debug. Instead,
simple DEBUG macros are used, like in GtkSourceView. To enable the
debugging, the code must be recompiled, unfortunately. Maybe in the
future we can use GLib functions for the debugging.