Adapter between GtkSourceFile and file saving gedit API
The GeditDocument API for file saving and loading is used by some
plugins, especially the loaded and saved signals. It is simpler to no
break the gedit plugins as a first step.

This commit removes GeditMountOperationFactory (which was used only in
private functions) and uses GtkSourceMountOperationFactory instead. The
first argument to the callback function has been removed (a pointer to a
GeditDocument). The userdata argument is sufficient.