Create gcode
With gcode-debug, a copy of gedit-debug.

The goal of gcode is to create a library to create text editors easily.
gcode will remain internal to gedit. The long-term goal is to move the
code to GtkSourceView. With gcode we don't need to pay attention to API
stability, we can do refactorings later to improve the APIs. But ideally
the gedit API (for plugins) must not be broken.