Move gcode/ inside gedit/
This will be needed for GeditDocument and other gedit public classes.
GeditDocument needs to be a subclass of GcodeDocument, so the
GcodeDocument and/or GcodeDocumentClass structs need to be public (for
gedit plugins). So the idea is to have one or more headers in gcode that
are installed with the other gedit headers. So those gcode headers need
to be in a 'gedit' directory.

So in gedit-document.h we could include a gcode public header like this:
 #include <gedit/gcode/gcode-gedit.h>

Another solution is to have an include like this:
 #include <gcode/gcode-blah.h>

But the latter would make the gcode namespace "public" for installed
headers in the system. It's better to keep it inside gedit.